midleware

+++ routes/api.php
@@ -37,31 +37,31 @@ Route::get('receta/{id}', [RecetaController::class, 'obtenerReceta']);
 Route::post('buscador', [RecetaController::class, 'buscarReceta']);
 Route::post('nuevaReceta', [RecetaController::class, 'crear']); // hecho
 Route::get('recetas/getAll', [RecetaController::class , 'getAll']);
-Route::post('modificarReceta', [RecetaController::class, 'modificarReceta']);
-Route::post('recetasPorCategoria', [RecetaController::class, 'recetasPorCategoria']);
-Route::post('recetasPorNutriscore', [RecetaController::class, 'recetasPorNutriscore']);
+Route::middleware('admin')->post('modificarReceta', [RecetaController::class, 'modificarReceta']);
+Route::middleware('admin')->post('recetasPorCategoria', [RecetaController::class, 'recetasPorCategoria']);
+Route::middleware('admin')->post('recetasPorNutriscore', [RecetaController::class, 'recetasPorNutriscore']);
 Route::get('actualizarPrecios', [RecetaController::class, 'actualizarPrecios']); // hecho
 
 //COMENTARIOS
 Route::get('comentariosUsuario/{id}', [ComentariosController::class, 'comentariosUsuario']);
-Route::post('nuevoComentario', [ComentariosController::class, 'nuevoComentario']);
+Route::middleware('auth')->post('nuevoComentario', [ComentariosController::class, 'nuevoComentario']);
 Route::get('comentariosReceta/{id}', [ComentariosController::class, 'comentariosReceta']);
-Route::get('numeroComentarios', [ComentariosController::class, 'numeroComentarios']);
+Route::middleware('admin')->get('numeroComentarios', [ComentariosController::class, 'numeroComentarios']);
 
 //USUARIOS
 Route::post('login', [UsuarioController::class, 'login']); // hecho
 Route::get('logout', [UsuarioController::class, 'logout']);
 Route::post('registro', [UsuarioController::class, 'registrar']); // hecho
 Route::get('infoUsuario/{id}', [UsuarioController::class, 'infoUsuario']);
-Route::post('eliminarUsuario', [UsuarioController::class, 'eliminarUsuario']);
-Route::get('numeroUsuarios', [UsuarioController::class, 'numeroUsuarios']); //hecho
+Route::middleware('admin')->post('eliminarUsuario', [UsuarioController::class, 'eliminarUsuario']);
+Route::middleware('admin')->get('numeroUsuarios', [UsuarioController::class, 'numeroUsuarios']); //hecho
 Route::get('refresh', [UsuarioController::class, 'refresh']); //hecho
 Route::post('me', [UsuarioController::class, 'me']); //hecho
 
 //FAVORITOS
-Route::get('addFavoritos/{id_receta}', [FavoritoController::class, 'addFavoritos']); // hecho
-Route::delete('removeFavoritos/{id_receta}',[FavoritoController::class, 'removeFavoritos']); // hecho
-Route::get('esFavorito/{id_receta}', [FavoritoController::class, 'esFavorito']); // hecho
+Route::middleware('auth')->get('addFavoritos/{id_receta}', [FavoritoController::class, 'addFavoritos']); // hecho
+Route::middleware('auth')->delete('removeFavoritos/{id_receta}',[FavoritoController::class, 'removeFavoritos']); // hecho
+Route::middleware('auth')->get('esFavorito/{id_receta}', [FavoritoController::class, 'esFavorito']); // hecho
 
 //INGREDIENTES
 Route::get('ingredientes', [IngredienteController::class, 'ingredientes']); // hecho
@@ -76,7 +76,7 @@ Route::post('nuevoPrecio', [PrecioController::class, 'crear']); // hecho
 
 
 //ADMIN
-Route::middleware('admin')->get('getColores', [OpcionesAdminController::class, 'getColores']); //hecho
+Route::get('getColores', [OpcionesAdminController::class, 'getColores']); //hecho
 Route::middleware('admin')->post('setPrincipal', [OpcionesAdminController::class, 'setPrincipal']); //hecho
 Route::middleware('admin')->post('setClaro', [OpcionesAdminController::class, 'setClaro']); //hecho
 Route::middleware('admin')->post('setSecundario', [OpcionesAdminController::class, 'setSecundario']); //hecho
